+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Japanese Kanji support added using [`kuromoji`](https://pub.dev/packages/kuromoji).
 - Hebrew romanization support added based on ISO 259 / DIN 31636.
+- Added support for multiple Arabic romanization systems (ALA-LC, DIN 31635, Buckwalter).
 
 ## 0.0.1